requirements.txt
Django==5.0.0
djangorestframework==3.15.0
python-docx==1.1.0
pdfminer.six==20231228
rapidfuzz==3.10.0
sentence-transformers==2.7.0
nltk==3.8.1
pytest==8.3.0
pytest-django==4.8.0
python-dotenv==1.0.1
Pillow==10.4.0
